Automatic Options Persistance
Added useOnce hook to call a function a single time from within a component.
Configured App component to load the options from the config file, if it exists, once.
Configured Options component to save the options whenever any of the values changes; It's not super efficient but this is much more intuitive than displaying an explicit Save button that the user must click. It just works.