Refining UI
Set up a minimum and default window size for the app.
Renamed SimpleInput to LabelInputRow.
Created InputRow, which is nearly the same as LabelInputRow only without the label element.
Added optional inputs for the title and placeholder properties on the InputRow and SimpleInputRow components.
Moved the #showOptions button into the same row as the video URL input to clean up the layout a bit; It retains its z-index so it can continue to be used to hide the Options component as well as show it.