Update video.rs
Created DownloadProgress struct to parse the necessary information from yt-dlp's stdout output and transport it back to a rust-vdl component.
Updated spawnCommand to include the --newline option so the download progress is printed on new lines instead of updating a single line; Makes it simpler to parse in real time.
Renamed writeOutput to processOutput