ProgressBar Update 1
Updated Cargo.toml to pull the latest master branch for all Dioxus dependencies, and using liberal version specifications for the rest. Added process feature to tokio dependency.

Created VideoDownloaderOptions struct to contain all the necessary CLI options to pass when executing the yt-dlp command.

Updated the Options component to allow editing of all the new options; Changed the minimum window size in constants.rs to account for the larger amount of options.

Updated the global state values, as well as the save/load methods for storing and retrieving user-configured options, in state.rs to utilize the new VideoDownloaderOptions struct.

Added a DownloadElement component to handle rendering the progress bar and controlling an individual download task. Currently only allows the user to halt and resume downloads. The UI is functional but is not yet final.

Updated VideoDownloader to use tokio::process versions of Command and its related types; Required marking the download and processOutput methods as async, as well as using tokio-stream to handle parsing the output, since tokio::process::ChildStdout/ChildStderr are AsyncRead rather than Read.