Update Progress Component
Moved the list of URLs to be downloaded into an AtomRef in state.rs.

Added a Remove button to the DownloadElement component that is displayed when a download is halted. Ensures the download task is aborted before removing the url from the UrlList.

Further refined the styles of the whole application.

Fixed options component not triggering scroll bars; Changed minimum window size to 500x500.