Resolving dead_code Warnings
Moved the three get*Dir functions out of state.rs into dir.rs.

Commented out the template module in download/mod.rs, as it is not being used at the moment. It will be used by a UI element that is planned but not actively being implemented yet.

Updated VideoDownloaderOptions to utilize getUserDownloadsDir() as the default outputPath value as well as actually using the Default_OutputTemplate constant for the default output value.