Updated Downloader
Added subLangs property to VideoDownloaderOptions and the Options component.

Updated DownloadProgress to be able to transmit a list of subtitle languages which have been found, based on the subLangs option.

Updated the DownloadProgress::toString method to be more robust, allowing for useful logs when a video title, video formats, or subtitle languages update is transmitted.